make submit_time.py skip solved and fast problems

+++ submit_time.py
@@ -24,23 +24,26 @@ def main():
             problems_list = []
             probIDs = [name for name in os.listdir('trainings/%s'%(size))]
         for problem in probIDs:
-            #we can do many of these at once if we so desire
-            startTime = time.time()
-            currentProcess = subprocess.Popen(["./make_guess",size,problem,"time"] + problems_list)
-            done = False
-            while ((time.time() - startTime) < 4*60) and (not done):
-                if (currentProcess.poll() == 0):
-                    with open("%s/%s/fast"%(dirname,problem),"w") as notefile:
+            fastname = "%s/%s/fast"%(dirname,problem)
+            solvdname = "%s/%s/solved"%(dirname,problem)
+            if not os.path.exists(fastname) and not os.path.exists(solvedname):
+                #we can do many of these at once if we so desire
+                startTime = time.time()
+                currentProcess = subprocess.Popen(["./make_guess",size,problem,"time"] + problems_list)
+                done = False
+                while ((time.time() - startTime) < 5*60) and (not done):
+                    if (currentProcess.poll() == 0):
+                        with open("%s/%s/fast"%(dirname,problem),"w") as notefile:
+                            pass
+                        done = True
+                    time.sleep(1)
+                try:
+                    currentProcess.kill()
+                    print "==== process did not finish ++++++++++++++++++++++++++++++++++++++++++++++++++"
+                    with open("%s/%s/slow"%(dirname,problem),"w") as notefile:
                         pass
-                    done = True
-                time.sleep(1)
-            try:
-                currentProcess.kill()
-                print "==== process did not finish ++++++++++++++++++++++++++++++++++++++++++++++++++"
-                with open("%s/%s/slow"%(dirname,problem),"w") as notefile:
-                    pass
-            except:
-                print "==== process finished --------------------------------------------------"
+                except:
+                    print "==== process finished --------------------------------------------------"
         return 0
 
 if __name__ == '__main__':
